+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 - Truncate decimal values in date input fields to just the integer component
 - Remove duplicate date values from date input fields (eg, '2,2,2,2' will result in '2')
 - Disable input fields that are not related to the sub-type selection for Monthly and Yearly options
+- When validating, throw an error when any of the provided days are larger than 31
 
 ## [1.0.0] - 2016-06-03
 - Extracted default template to seperate file to make editing easier.